+++ src/browser_actions/browser-action.js
@@ -1,181 +1,17 @@
 import moment from 'moment';
+import * as Sidebar from './util/sidebar';
+import * as MangaList from './util/list';
 import * as Notification from '../util/notification';
 import store from '../util/datastore';
 
-
-// Document startup and event listeners
-// ////////////////////////////////////////////////////////////////
-
 const mangaListDom = document.getElementById('manga-list');
 
-/**
- * Opens new tab with the selected manga chapter when btn-read button is clicked.
- */
-async function readButtonListener(e) {
-  try {
-    const select = e.target.parentElement.getElementsByTagName('select')[0];
-    await browser.tabs.create({ url: select.options[select.selectedIndex].dataset.url });
-  } catch (err) {
-    console.error(`Error while accessing chapter page: ${err}`); // eslint-disable-line no-console
 
-    // Notify user that an error occurred
-    Notification.error({
-      title: browser.i18n.getMessage('readChapterButtonErrorNotificationTitle'),
-      message: browser.i18n.getMessage('readChapterButtonErrorNotificationMessage'),
-    });
-  }
-}
-
-/**
- * Unbookmarks manga when btn-remove button is clicked.
- */
-async function unbookmarkButtonListener(e) {
-  const card = e.target.parentElement.parentElement;
-
-  try {
-    const result = await browser.runtime.sendMessage({
-      type: 'unbookmark',
-      manga_url: card.dataset.mangaUrl,
-    });
-
-    if (result) mangaListDom.removeChild(card);
-  } catch (err) {
-    console.error(`Error while unbookmarking manga from browser action script: ${err}`); // eslint-disable-line no-console
-
-    // Notify user that an error occurred
-    Notification.error({
-      title: browser.i18n.getMessage('unbookmarkMangaErrorNotificationTitle'),
-      message: browser.i18n.getMessage('unbookmarkMangaErrorNotificationmessage'),
-    });
-  }
-}
-
-/**
- * Listens to 'click' events on options-btn. Opens new tab with the selected manga chapter when btn-read button is clicked.
- */
-document.addEventListener('click', async (e) => {
-  if (e.target.id !== 'options-btn') return;
-
-  try {
-    await browser.runtime.openOptionsPage();
-  } catch (err) {
-    console.error(`Error while opening options page: ${err}`); // eslint-disable-line no-console
-
-    // Notify user that an error occurred
-    Notification.error({
-      title: browser.i18n.getMessage('optionsPageButtonErrorNotificationTitle'),
-      message: browser.i18n.getMessage('optionsPageButtonErrorNotificationMessage'),
-    });
-  }
-});
-
-/**
- * Listens to 'change' events on the view mode container.
- */
-document.addEventListener('change', async (e) => {
-  if (e.target.name !== 'viewMode') return;
-
-  try {
-    await browser.storage.sync.set({ view_mode: e.target.id });
-
-    const container = document.getElementById('view-mode-container');
-    const previousActive = container.getElementsByClassName('active')[0];
-
-    if (previousActive) previousActive.classList.remove('active');
-
-    e.target.parentNode.classList.add('active');
-  } catch (err) {
-    console.error(`Could not change the view mode: ${err}`); // eslint-disable-line no-console
-
-    // Notify user that an error occurred
-    Notification.error({
-      title: browser.i18n.getMessage('changeViewModeErrorNotificationTitle'),
-      message: browser.i18n.getMessage('changeViewModeErrorNotificationMessage'),
-    });
-  }
-});
-
-
-// DOM creation
+// Event listeners
 // ////////////////////////////////////////////////////////////////
 
-/**
- * Creates a new Manga <li> Element
- * @param {*} manga object defining the manga
- * @returns <li> html element
- */
-function createMangaElement(bookmark, manga) {
-  const upToDate = bookmark.last_read.chapter.id === manga.chapter_list[manga.chapter_list.length - 1].id;
-
-  const mangaDiv = document.createElement('li');
-  mangaDiv.id = `${bookmark.source}-${bookmark.reference}`;
-  mangaDiv.className = 'manga-row flex';
-  mangaDiv.dataset.mangaUrl = manga.url;
-  mangaDiv.dataset.mangaSid = manga.sid;
-
-  if (upToDate) mangaDiv.className += ' up-to-date';
-
-  // Add cover image
-  const mangaCover = document.createElement('img');
-  mangaCover.src = manga.cover;
-  mangaCover.width = '100';
-
-  mangaDiv.appendChild(mangaCover);
-
-  const mangaData = document.createElement('div');
-  mangaData.className = 'manga-data';
-
-  // Add Manga title
-  const title = document.createElement('h4');
-  title.textContent = manga.name;
-  title.className = 'title';
-  mangaData.appendChild(title);
-
-  // Add Manga metadata
-  const meta = document.createElement('span');
-  meta.textContent = `Last read: ${moment(bookmark.last_read.date).format('LL')}`;
-  mangaData.appendChild(meta);
-
-  // Add chapters
-  const chapterSelect = document.createElement('select');
-  mangaData.appendChild(chapterSelect);
-  if (upToDate) {
-    chapterSelect.className = 'chapter-selector bg-success';
-  } else {
-    chapterSelect.className = 'chapter-selector bg-danger';
-  }
-
-  manga.chapter_list.slice().reverse().forEach((chapter) => {
-    const option = document.createElement('option');
-    option.value = chapter.id;
-    option.text = chapter.name;
-    option.dataset.url = chapter.url;
-
-    if (chapter.id === bookmark.last_read.chapter.id) option.selected = true;
-
-    chapterSelect.appendChild(option);
-  });
-
-  // Add Read button
-  const goBtn = document.createElement('a');
-  goBtn.className = 'btn btn-default btn-read';
-  goBtn.innerHTML = 'Read';
-  goBtn.href = '#';
-  goBtn.onclick = readButtonListener;
-  mangaData.appendChild(goBtn);
-
-  // Add Remove button
-  const removeBtn = document.createElement('a');
-  removeBtn.className = 'btn btn-danger btn-remove';
-  removeBtn.innerHTML = 'Remove';
-  removeBtn.href = '#';
-  removeBtn.onclick = unbookmarkButtonListener;
-  mangaData.appendChild(removeBtn);
-
-  mangaDiv.appendChild(mangaData);
-
-  return mangaDiv;
-}
+document.addEventListener('click', Sidebar.expandButtonListener);
+document.addEventListener('change', Sidebar.viewModeListener);
 
 
 // Runtime messages
@@ -193,21 +29,25 @@ function updateCurrentChapter(bookmark) {
 
   const uptodate = bookmark.last_read.chapter.index === chapterSel.options.length - 1;
   if (uptodate) {
-    chapterSel.className = 'chapter-selector bg-success';
-    mangaDom.className += ' up-to-date';
+    chapterSel.classList.replace('bg-danger', 'bg-success');
+    mangaDom.classList.add('flex-last');
+    Sidebar.updateChart(-1, 1);
   }
 }
 
 function updateChapterList(bookmark, chapterList) {
   const mangaDom = document.getElementById(`${bookmark.source}-${bookmark.reference}`);
-  const wasUptodate = mangaDom.className.includes('up-to-date');
+  const wasUptodate = mangaDom.classList.contains('flex-last');
 
-  if (wasUptodate) mangaDom.className = 'manga-row flex';
+  const chapterSel = mangaDom.getElementsByTagName('select')[0];
+
+  if (wasUptodate) {
+    mangaDom.classList.remove('flex-last');
+    chapterSel.classList.replace('bg-success', 'bg-danger');
+    Sidebar.updateChart(1, -1);
+  }
 
   // Modify chapters
-  const chapterSel = mangaDom.getElementsByTagName('select')[0];
-  if (wasUptodate) chapterSel.className = 'chapter-selector bg-danger';
-
   chapterSel.options.length = 0;
   chapterList.slice().reverse().forEach((chapter) => {
     const option = document.createElement('option');
@@ -250,11 +90,20 @@ window.onload = async () => {
     // Sanity check the DOM
     if (!mangaListDom) throw new Error('manga-list element does not exist in DOM');
 
-    const storage = await browser.storage.sync.get(['bookmark_list', 'badge_count', 'view_mode']);
+    const storage = await browser.storage.sync.get();
+
+    // Initialize sidebar
+    Sidebar.init({
+      last_update: storage.last_chapter_update,
+    });
 
     // Select view mode
     const radio = document.getElementById(storage.view_mode || 'single');
     if (radio) radio.parentNode.className += ' active';
+
+    // Select sidebar mode
+    const iconBtn = document.getElementById('sidebar-expand-btn');
+    if (storage.sidebar_collapsed) Sidebar.collapse(iconBtn.firstElementChild);
 
     // Update badge_count
     if (storage && storage.badge_count > 0) {
@@ -266,16 +115,25 @@ window.onload = async () => {
     }
 
     // console.debug(storage);
+    let readCount = 0;
 
     // Append manga list
     if (storage && storage.bookmark_list) {
-      storage.bookmark_list.forEach(async (bookmark) => {
+      const promises = storage.bookmark_list.map(async (bookmark) => {
         const manga = await store.getItem(`${bookmark.source}/${bookmark.reference}`);
         if (!manga) return;
 
-        mangaListDom.appendChild(createMangaElement(bookmark, manga));
+        const card = MangaList.createMangaCard(bookmark, manga);
+        mangaListDom.appendChild(card);
+
+        if (card.classList.contains('flex-last')) readCount += 1;
       });
+
+      await Promise.all(promises);
     }
+
+    // Update chart
+    Sidebar.updateChart(mangaListDom.childElementCount - readCount, readCount);
   } catch (err) {
     console.error(`Error while starting the browser action script: ${err}`); // eslint-disable-line no-console